+++ example/advanced/server/advanced_server.cpp
@@ -621,6 +621,10 @@ public:
         if(ec && ec != boost::asio::error::operation_aborted)
             return fail(ec, "timer");
 
+        // Check if this has been upgraded to Websocket
+        if(timer_.expires_at() == (std::chrono::steady_clock::time_point::min)())
+            return;
+
         // Verify that the timer really expired since the deadline may have moved.
         if(timer_.expiry() <= std::chrono::steady_clock::now())
         {
@@ -658,6 +662,10 @@ public:
         // See if it is a WebSocket Upgrade
         if(websocket::is_upgrade(req_))
         {
+            // Make timer expire immediately, by setting expiry to time_point::min we can detect
+            // the upgrade to websocket in the timer handler
+            timer_.expires_at((std::chrono::steady_clock::time_point::min)());
+
             // Create a WebSocket websocket_session by transferring the socket
             std::make_shared<websocket_session>(
                 std::move(socket_))->do_accept(std::move(req_));